+++ packages/botonic-react/CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to Botonic will be documented in this file.
 ### Added
 
 - [PR-3127](https://github.com/hubtype/botonic/pull/3127): Added new system events in webchat preview.
+- [PR-3128](https://github.com/hubtype/botonic/pull/3128): Add new message Whatsapp Interactive Media Carousel.
 
 ### Changed